+++ server/__tests__/config-test.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import _ from "lodash";
 
 jest.dontMock("lodash");
@@ -12,14 +13,14 @@ const ship = {
 describe("buildConfigFromShip", () => {
   it("build a config object from the ship config object", () => {
 
-    const config = require("../config");
-    const getServiceAttributeToHullTopLevel = require("../mapping-data").getServiceAttributeToHullTopLevel;
-    const getServiceAttributeToHullTrait = require("../mapping-data").getServiceAttributeToHullTrait;
+    const config = require("../lib/config");
+    const getServiceAttributeToHullTopLevel = require("../lib/mapping-data").getServiceAttributeToHullTopLevel;
+    const getServiceAttributeToHullTrait = require("../lib/mapping-data").getServiceAttributeToHullTrait;
 
     const c = config.buildConfigFromShip(ship, organization, secret);
-    expect(_.keys(c)).toEqual(["hull", "salesforce", "sync", "mappings"]);
-    expect(_.keys(c.mappings)).toEqual(["Lead", "Contact"]);
-    ["Lead", "Contact"].forEach((type) => {
+    expect(_.keys(c)).toEqual(["hull", "settings", "salesforce", "sync", "mappings"]);
+    expect(_.keys(c.mappings)).toEqual(["Lead", "Contact", "Account"]);
+    _.keys(c.mappings).forEach((type) => {
       expect(_.keys(c.mappings[type])).toEqual(["type", "fetchFields", "fields", "fetchFieldsToTopLevel"]);
       expect(c.mappings[type].fetchFields).toEqual(getServiceAttributeToHullTrait(type));
       expect(c.mappings[type].fetchFieldsToTopLevel).toEqual(getServiceAttributeToHullTopLevel(type));