+++ server/index.js
@@ -1,9 +1,28 @@
+import express from "express";
 import Hull from "hull";
-import { Agent } from "./agent";
-import { Server } from "./server";
-import config from "./config";
+import server from "./server";
 
-Hull.logger.transports.console.json = true;
-Hull.logger.info("start", { transport: "console" });
+if (process.env.NEW_RELIC_LICENSE_KEY) {
+  Hull.logger.warn("starting newrelic with key: ", process.env.NEW_RELIC_LICENSE_KEY);
+  // eslint-disable-next-line global-require
+  require("newrelic");
+}
 
-export default { Agent, Server, config };
+const config = {
+  hostSecret: process.env.SECRET || "BOOM",
+  port: process.env.PORT || 8082,
+  salesforce: {
+    oauth2: {
+      clientId: process.env.CLIENT_ID,
+      clientSecret: process.env.CLIENT_SECRET
+    }
+  },
+};
+
+const connector = new Hull.Connector(config);
+const app = express();
+
+connector.setupApp(app);
+
+server(app, config);
+connector.startApp(app);