+++ server/lib/mapping-data.js
@@ -24,13 +24,13 @@ const DEFAULT_MAPPING = {
     { service_attribute: "LastActivityDate", hull_top_level_trait: null, hull_trait: "last_activity_date", type: "string" },
 
     { service_attribute: "ConvertedDate", hull_top_level_trait: null, hull_trait: "converted_at", type: "string" },
-    
+
     // Address object
     { service_attribute: "City", hull_top_level_trait: "address.city", hull_trait: "city", type: "string" },
     { service_attribute: "PostalCode", hull_top_level_trait: "address.postal_code", hull_trait: "postal_code", type: "string" },
     { service_attribute: "State", hull_top_level_trait: "address.state", hull_trait: "state", type: "string" },
     { service_attribute: "Country", hull_top_level_trait: "address.country", hull_trait: "country", type: "string" },
-    
+
     // { service_attribute: "MiddleName", hull_top_level_trait: null, hull_trait: "middle_name", type: "string" }, // No such column 'MiddleName' on entity 'Lead'
     { service_attribute: "Industry", hull_top_level_trait: null, hull_trait: "industry", type: "string" },
     // { service_attribute: "CountryCode", hull_top_level_trait: null, hull_trait: "country_code", type: "string" }, // No such column 'CountryCode' on entity 'Lead'
@@ -62,16 +62,30 @@ const DEFAULT_MAPPING = {
     { service_attribute: "MailingPostalCode", hull_top_level_trait: "address.postal_code", hull_trait: "mailing_postal_code", type: "string" },
     { service_attribute: "MailingState", hull_top_level_trait: "address.state", hull_trait: "mailing_state", type: "string" },
     { service_attribute: "MailingCountry", hull_top_level_trait: "address.country", hull_trait: "mailing_country", type: "string" },
-    
+
     // { service_attribute: "MiddleName", hull_top_level_trait: null, hull_trait: "middle_name", type: "string" }, // No such column 'MiddleName' on entity 'Contact'
     { service_attribute: "Birthdate", hull_top_level_trait: null, hull_trait: "birthdate", type: "string" },
     // { service_attribute: "DoNotCall", hull_top_level_trait: null, hull_trait: "do_not_call", type: "string" }, // No such column 'DoNotCall' on entity 'Contact'
     // { service_attribute: "MailingCountryCode", hull_top_level_trait: null, hull_trait: "mailing_country_code", type: "string" }, // No such column 'MailingCountryCode' on entity 'Contact'
     // { service_attribute: "Website", hull_top_level_trait: "website", hull_trait: "website", type: "string" }, // No such column 'Website' on entity 'Contact'
     { service_attribute: "Id", hull_top_level_trait: null, hull_trait: "id", type: "string" },
-    { service_attribute: "OwnerId", hull_top_level_trait: null, hull_trait: "owner_id", type: "string" }
+    { service_attribute: "OwnerId", hull_top_level_trait: null, hull_trait: "owner_id", type: "string" },
+    { service_attribute: "AccountId", hull_top_level_trait: null, hull_trait: "account_id", type: "string" },
+  ],
+  Account: [
+    { service_attribute: "Id", hull_top_level_trait: null, hull_trait: "id", type: "string" },
+    { service_attribute: "AnnualRevenue", hull_top_level_trait: null, hull_trait: "annual_revenue", type: "string" },
+    { service_attribute: "NumberOfEmployees", hull_top_level_trait: null, hull_trait: "number_of_employees", type: "int" },
+    { service_attribute: "Industry", hull_top_level_trait: null, hull_trait: "industry", type: "string" },
+    { service_attribute: "Name", hull_top_level_trait: "name", hull_trait: "name", type: "string" },
+    { service_attribute: "Phone", hull_top_level_trait: null, hull_trait: "phone", type: "string" },
+    { service_attribute: "Website", hull_top_level_trait: "domain", hull_trait: "website", type: "string" }
   ]
 };
+
+export function getTypes() {
+  return _.keys(DEFAULT_MAPPING);
+}
 
 /**
  * Returns a mapping between Salesforce attributes and Hull top level