+++ server/lib/sync.js
@@ -1,7 +1,9 @@
 import Hogan from "hogan.js";
 import _ from "lodash";
 
-export function getUpdatedFields(user, sfObject, mapping) {
+// NOTE: exported for tests only
+export function getUpdatedFields(user, sfObject, mapping, initialRecord) {
+  const initialSize = Object.keys(initialRecord).length;
   const fields = mapping.fields;
   const fieldNames = Object.keys(fields);
   const record = fieldNames.reduce((mapped, f) => {
@@ -15,9 +17,12 @@ export function getUpdatedFields(user, sfObject, mapping) {
 
     if (orig === undefined || orig === null || orig === (def && def.defaultValue) || def.overwrite === true) {
       if (typeof (def) === "string") {
-        val = user[def];
+        const key = _.last(_.split(def, "."));
+        val = user[key];
       } else if (def.key) {
-        val = _.get(user, def.key);
+        // Quick fix to handle accounts fields that are defined as `account.*`
+        const key = _.last(_.split(def.key, "."));
+        val = _.get(user, key);
       } else if (def.tpl) {
         val = Hogan.compile(def.tpl).render(user);
       }
@@ -34,15 +39,15 @@ export function getUpdatedFields(user, sfObject, mapping) {
     }
 
     return mapped;
-  }, { Email: user.email });
+  }, initialRecord);
 
-  if (Object.keys(record).length > 1) {
+  if (Object.keys(record).length > initialSize) {
     return record;
   }
   return undefined;
 }
 
-export function syncRecords(sfObjectsByEmail, users, options) {
+export function syncUsers(sfObjectsByEmail, users, options) {
   return users.reduce((records, user) => {
     const sfObjects = sfObjectsByEmail[user.email] || {};
 
@@ -51,7 +56,7 @@ export function syncRecords(sfObjectsByEmail, users, options) {
     const mapping = options.mappings[objectType];
 
     if (mapping) {
-      const record = getUpdatedFields(user, sfObjects[objectType], mapping);
+      const record = getUpdatedFields(user, sfObjects[objectType], mapping, { Email: user.email });
       if (record) {
         records[objectType].push(record);
       }
@@ -59,4 +64,26 @@ export function syncRecords(sfObjectsByEmail, users, options) {
 
     return records;
   }, { Contact: [], Lead: [] });
+}
+
+/*
+ * :param sfAccounts: object - Salesforce Account records by domain or id
+ */
+export function syncAccounts(sfAccounts, accounts, mapping) {
+  return accounts.reduce((records, account) => {
+    const sfAccount = sfAccounts[account.domain] || sfAccounts[account["salesforce/id"]] || {};
+
+    const accountToUpsert = {};
+    if (sfAccount.Id) {
+      accountToUpsert.Id = sfAccount.Id;
+    }
+
+    const record = getUpdatedFields(account, sfAccount, mapping, accountToUpsert);
+
+    if (record) {
+      records.push(record);
+    }
+
+    return records;
+  }, []);
 }