+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.4.2
+- [feature] use account salesforce id to sync when it is available in hull
+
+## v0.4.1
+- [feature] export hull accounts to salesforce matching domain to Website
+
+## v0.4.0
+- [feature] import salesforce accounts into hull
+
 ## v0.3.0
 - [feature] do not fetch all users when no synchronized segment is defined
 - [feature] fetch all default salesforce attributes